+++ packages/shelf_api_builder/lib/src/analyzers/methods_analyzer.dart
@@ -4,6 +4,7 @@ import 'package:meta/meta.dart';
 
 import '../models/endpoint_method.dart';
 import '../readers/api_method_reader.dart';
+import '../readers/shelf_api_reader.dart';
 import 'body_analyzer.dart';
 import 'path_analyzer.dart';
 import 'query_analyzer.dart';
@@ -22,23 +23,29 @@ class MethodsAnalyzer {
       _queryAnalyzer = QueryAnalyzer(buildStep),
       _responseAnalyzer = ResponseAnalyzer(buildStep);
 
-  Future<List<EndpointMethod>> analyzeMethods(ClassElement clazz) =>
-      _analyzeMethods(clazz).toList();
+  Future<List<EndpointMethod>> analyzeMethods(
+    ClassElement clazz,
+    ShelfApiReader shelfApi,
+  ) => _analyzeMethods(clazz, shelfApi).toList();
 
-  Stream<EndpointMethod> _analyzeMethods(ClassElement clazz) async* {
+  Stream<EndpointMethod> _analyzeMethods(
+    ClassElement clazz,
+    ShelfApiReader shelfApi,
+  ) async* {
     for (final method in clazz.methods) {
       final apiMethod = method.apiMethodAnnotation;
       if (apiMethod == null) {
         continue;
       }
 
-      yield await _analyzeMethod(method, apiMethod);
+      yield await _analyzeMethod(method, apiMethod, shelfApi);
     }
   }
 
   Future<EndpointMethod> _analyzeMethod(
     MethodElement method,
     ApiMethodReader apiMethod,
+    ShelfApiReader shelfApi,
   ) async => EndpointMethod(
     name: method.name!,
     httpMethod: apiMethod.method,
@@ -46,6 +53,10 @@ class MethodsAnalyzer {
     pathParameters: await _pathAnalyzer.analyzePath(method, apiMethod),
     body: await _bodyAnalyzer.analyzeBody(method),
     queryParameters: await _queryAnalyzer.analyzeQuery(method),
-    response: await _responseAnalyzer.analyzeResponse(method, apiMethod),
+    response: await _responseAnalyzer.analyzeResponse(
+      method,
+      apiMethod,
+      shelfApi,
+    ),
   );
 }