+++ packages/shelf_api_builder/lib/src/readers/api_method_reader.dart
@@ -23,6 +23,8 @@ class ApiMethodReader with SerializableReader {
   String get method => constantReader.read('method').stringValue;
 
   String get path => constantReader.read('path').stringValue;
+
+  bool? get autoNotFound => constantReader.peek('autoNotFound')?.boolValue;
 }
 
 @internal