+++ packages/shelf_api/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.4-dev.2] - 2026-05-26
+### Added
+- Added annotations for automatic not found handling
+
 ## [1.4.4-dev.1] - 2026-05-25
 ### Changed
 - Updated min sdk version to ^3.12.0
@@ -90,6 +94,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial Release
 
+[1.4.4-dev.2]: https://github.com/Skycoder42/shelf_api/compare/shelf_api-v1.4.4-dev.1...shelf_api-v1.4.4-dev.2
 [1.4.4-dev.1]: https://github.com/Skycoder42/shelf_api/compare/shelf_api-v1.4.3...shelf_api-v1.4.4-dev.1
 [1.4.3]: https://github.com/Skycoder42/shelf_api/compare/shelf_api-v1.4.2...shelf_api-v1.4.3
 [1.4.2]: https://github.com/Skycoder42/shelf_api/compare/shelf_api-v1.4.1...shelf_api-v1.4.2