+++ packages/shelf_api_builder/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.4-dev.2] - 2026-05-26
+### Added
+- Added optional automatic not found handling
+  - will convert `null` responses from handlers to 404 Not Found responses and back in the client
+  - Disabled by default
+
 ## [1.3.4-dev.1] - 2026-05-25
 ### Changed
 - Updated min sdk version to ^3.12.0
@@ -87,6 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
+[1.3.4-dev.2]: https://github.com/Skycoder42/shelf_api/compare/shelf_api_builder-v1.3.4-dev.1...shelf_api_builder-v1.3.4-dev.2
 [1.3.4-dev.1]: https://github.com/Skycoder42/shelf_api/compare/shelf_api_builder-v1.3.3...shelf_api_builder-v1.3.4-dev.1
 [1.3.3]: https://github.com/Skycoder42/shelf_api/compare/shelf_api_builder-v1.3.2...shelf_api_builder-v1.3.3
 [1.3.2]: https://github.com/Skycoder42/shelf_api/compare/shelf_api_builder-v1.3.1...shelf_api_builder-v1.3.2